Forwarder script is added to AshScriptPlugin
